chore(web-refactor): state uplifted for drive container in web component
